+++ daemons/hw_alsa_usb.c
@@ -161,7 +161,7 @@ static char *rec(struct ir_remote *remotes)
 		current.tv_sec - last_time.tv_sec <= 2 &&
 		time_elapsed(&last_time, &current) <= 350000;
 	last_time = current;
-	LOGPRINTF(1, "code: %llx", (unsigned long long) code);
+	LOGPRINTF(1, "code: %llx", (__u64) code);
 	LOGPRINTF(1, "repeat_flag: %d", repeat_flag);
 	return decode_all(remotes);
 }
